refactor: AccompanyReview 관련 DTO 패키지 구조 변경
- AccompanyReview 관련 DTO 클래스를 response 패키지로 이동했습니다.
- 더 명확한 패키지 구조로 코드 가독성을 개선했습니다.
