+++ src/main/java/connectripbe/connectrip_be/review/dto/response/AccompanyReviewListResponse.java
@@ -1,5 +1,6 @@
-package connectripbe.connectrip_be.review.dto;
+package connectripbe.connectrip_be.review.dto.response;
 
+import connectripbe.connectrip_be.review.dto.AccompanyReviewDto;
 import java.util.List;
 
 public record AccompanyReviewListResponse(
